Introduce operator class in Apollo
